+++ tests/workflows/test_hashing.py
@@ -5,7 +5,7 @@ from pharmpy.workflows.hashing import DatasetHash, ModelHash
 def test_hash(load_example_model_for_test):
     model = load_example_model_for_test("pheno")
     h = ModelHash(model)
-    assert str(h) == "v6Cc3F9e4r1F51dKitKIhLC-7qIZw8QBLFExhm_KNps"
+    assert str(h) == "v6dmvSNgacGyaoE4_nf42Y0LOvoYKlJI9TdxujnBjBU"
     d = DatasetHash(model.dataset)
     assert str(d) == h.dataset_hash